requirements.txt
pywebview
pymongo[srv]
python-dotenv
qtpy
PySide6
